requirements.txt
Flask
scikit-learn
numpy
joblib
openai
python-dotenv